[feat] 새로운 기록물 만들기 기능추가
이제 나만의 새로운 기록물을 만들 수 있어요.
markdown형태의 글을 작성하고 저장할 수 있어요. (사진/동영상/임베드 등 기능 X)

추후에 태그 기능을 포함하여 write(새로운 기록물 만들기) 기능들을 추가해나갈 예정이에요.

+++ src/app/api/delete-article/route.ts
@@ -25,8 +25,19 @@ export async function DELETE(request: Request) {
             return new Response('Invalid _id parameter.', { status: 400 });
         }
 
+        const article = await db.collection('article').findOne({_id:objectId});
 
-        // Delete the document with the specified tagname
+        if (!article) {
+            return new Response('Article is already deleted', { status: 400 });
+        }
+
+        if(article.flatform == 'scraping'){
+            const url = article.url;
+            const articleId = url.match(/article\/([a-f0-9]{24})/)[1];
+
+            const result = await db.collection('write').deleteOne({_id: new ObjectId(articleId)});
+        }
+
         const result = await db.collection('article').deleteOne({ _id:objectId });
 
         if (result.deletedCount === 0) {
